Feat: 레인의 특성 로직 추가. 피유의 Ex, 패시브, 특성 로직 추가.
 - float로 계산되어야 작동하는 일부 보너스 스탯 계산식에 형변환 추가
 - 새 디버프: 부식 / 표적
 - 현재 코스트 관련 로직은 미구현 상태임.

+++ Assets/0-Scripts/Objects/Characters/Passive/AbilityLogics/Ability_322.cs
@@ -28,17 +28,17 @@ public class Ability_322: PassiveLogic
 
         if (healthPercentage <= coefficient["hpRatio_1"])
         {
-            stat_322.ReducedDmg = coefficient["reducedDmgRate_1"] / coefficient["denominator"];
+            stat_322.ReducedDmg = (float)(coefficient["reducedDmgRate_1"]) / coefficient["denominator"];
             character.tempBonusStat.AddBonusStat(stat_322);
         }
         else if (healthPercentage <= coefficient["hpRatio_2"])
         {
-            stat_322.ReducedDmg = coefficient["reducedDmgRate_2"] / coefficient["denominator"];
+            stat_322.ReducedDmg = (float)(coefficient["reducedDmgRate_2"]) / coefficient["denominator"];
             character.tempBonusStat.AddBonusStat(stat_322);
         }
         else if (healthPercentage <= coefficient["hpRatio_3"])
         {
-            stat_322.ReducedDmg = coefficient["reducedDmgRate_3"] / coefficient["denominator"];
+            stat_322.ReducedDmg = (float)(coefficient["reducedDmgRate_3"]) / coefficient["denominator"];
             character.tempBonusStat.AddBonusStat(stat_322);
         }
         else
